Overwrite the visit_docinfo and depart_docinfo methods to pass
In order to allow the user to specify the date on the title of the
presentation, bibliographic data and the --use-latex-docinfo flag
should be used.  However, this leads to the LaTeXTranslator (from
which BeamerTranslator inherits) adding a table that should include
the docinfo.  In rst2beamer, the table is empty and this just leads to
a blank slide.  So, I overwrote the visit_docinfo and depart_docinfo
methods to do nothing (just pass).  This seems to have eliminated the
empty table of docinfo and the resulting blank slide.  The date and
author and such go on the first slide thanks to \maketitle.